Moved addesslist components
